resize2fs: Fix up to be 64-bit block number safe

+++ resize/extent.c
@@ -19,22 +19,22 @@
 #include "resize2fs.h"
 
 struct ext2_extent_entry {
-	__u32	old_loc, new_loc;
-	int	size;
+	__u64	old_loc, new_loc;
+	__u64	size;
 };
 
 struct _ext2_extent {
 	struct ext2_extent_entry *list;
-	int	cursor;
-	int	size;
-	int	num;
-	int	sorted;
+	__u64	cursor;
+	__u64	size;
+	__u64	num;
+	__u64	sorted;
 };
 
 /*
  * Create an extent table
  */
-errcode_t ext2fs_create_extent_table(ext2_extent *ret_extent, int size)
+errcode_t ext2fs_create_extent_table(ext2_extent *ret_extent, __u64 size)
 {
 	ext2_extent	extent;
 	errcode_t	retval;
@@ -77,12 +77,12 @@ void ext2fs_free_extent_table(ext2_extent extent)
 /*
  * Add an entry to the extent table
  */
-errcode_t ext2fs_add_extent_entry(ext2_extent extent, __u32 old_loc, __u32 new_loc)
+errcode_t ext2fs_add_extent_entry(ext2_extent extent, __u64 old_loc, __u64 new_loc)
 {
 	struct	ext2_extent_entry	*ent;
 	errcode_t			retval;
-	int				newsize;
-	int				curr;
+	__u64				newsize;
+	__u64				curr;
 
 	if (extent->num >= extent->size) {
 		newsize = extent->size + 100;
@@ -139,10 +139,10 @@ static EXT2_QSORT_TYPE extent_cmp(const void *a, const void *b)
  * Given an inode map and inode number, look up the old inode number
  * and return the new inode number.
  */
-__u32 ext2fs_extent_translate(ext2_extent extent, __u32 old_loc)
+__u64 ext2fs_extent_translate(ext2_extent extent, __u64 old_loc)
 {
-	int	low, high, mid;
-	__u32	lowval, highval;
+	__s64	low, high, mid;
+	__u64	lowval, highval;
 	float	range;
 
 	if (!extent->sorted) {
@@ -170,7 +170,7 @@ __u32 ext2fs_extent_translate(ext2_extent extent, __u32 old_loc)
 			else
 				range = ((float) (old_loc - lowval)) /
 					(highval - lowval);
-			mid = low + ((int) (range * (high-low)));
+			mid = low + ((__u64) (range * (high-low)));
 		}
 #endif
 		if ((old_loc >= extent->list[mid].old_loc) &&
@@ -190,14 +190,14 @@ __u32 ext2fs_extent_translate(ext2_extent extent, __u32 old_loc)
  */
 void ext2fs_extent_dump(ext2_extent extent, FILE *out)
 {
-	int	i;
+	__u64	i;
 	struct ext2_extent_entry *ent;
 
 	fputs(_("# Extent dump:\n"), out);
-	fprintf(out, _("#\tNum=%d, Size=%d, Cursor=%d, Sorted=%d\n"),
+	fprintf(out, _("#\tNum=%llu, Size=%llu, Cursor=%llu, Sorted=%llu\n"),
 	       extent->num, extent->size, extent->cursor, extent->sorted);
 	for (i=0, ent=extent->list; i < extent->num; i++, ent++) {
-		fprintf(out, _("#\t\t %u -> %u (%d)\n"), ent->old_loc,
+		fprintf(out, _("#\t\t %llu -> %llu (%llu)\n"), ent->old_loc,
 			ent->new_loc, ent->size);
 	}
 }
@@ -205,8 +205,8 @@ void ext2fs_extent_dump(ext2_extent extent, FILE *out)
 /*
  * Iterate over the contents of the extent table
  */
-errcode_t ext2fs_iterate_extent(ext2_extent extent, __u32 *old_loc,
-				__u32 *new_loc, int *size)
+errcode_t ext2fs_iterate_extent(ext2_extent extent, __u64 *old_loc,
+				__u64 *new_loc, __u64 *size)
 {
 	struct ext2_extent_entry *ent;
